requirements.txt
# Cam_detect - Analyse vidéo de circulation
# Dépendances Python pour détection d'objets et analyse de trafic

# Computer Vision et détection d'objets
ultralytics>=8.0.0      # YOLOv8 - meilleure performance pour véhicules (95.5% mAP)
opencv-python>=4.8.0    # Traitement vidéo et images
torch>=2.0.0           # PyTorch pour les modèles de deep learning
torchvision>=0.15.0    # Vision utilities pour PyTorch
numpy>=1.24.0          # Calculs numériques
pillow>=10.0.0         # Manipulation d'images

# Interface graphique
# tkinter               # Interface utilisateur (inclus avec Python)
customtkinter>=5.2.0   # Interface moderne pour Windows 11
matplotlib>=3.7.0      # Graphiques et visualisations
seaborn>=0.12.0        # Visualisations statistiques avancées

# Traitement de données et analyse
pandas>=2.0.0          # Manipulation de données structurées
scipy>=1.10.0          # Calculs scientifiques
scikit-learn>=1.3.0    # Machine learning pour classification avancée

# Vidéo et multimédia
moviepy>=1.0.3         # Édition et traitement vidéo
imageio>=2.31.0        # I/O pour images et vidéos
imageio-ffmpeg>=0.4.8  # Support FFmpeg

# Utilitaires système
psutil>=5.9.0          # Monitoring système et performance
# pathlib              # Gestion des chemins (inclus avec Python 3.4+)
# json                 # Sérialisation de données (inclus avec Python)
# datetime             # Gestion des dates/temps (inclus avec Python)

# Configuration et logs
pyyaml>=6.0           # Configuration files
loguru>=0.7.0         # Système de logging avancé

# Performance et optimisation
numba>=0.57.0         # Accélération JIT pour calculs numériques
# threading           # Multi-threading (inclus avec Python)

# Base de données (optionnel pour historique)
# sqlite3             # Base de données locale (inclus avec Python)

# Tests et développement (optionnel)
pytest>=7.4.0        # Framework de tests
pytest-cov>=4.1.0    # Coverage des tests